+++ src/main/java/top/theillusivec4/curiouselytra/loader/CuriousElytraMod.java
@@ -21,19 +21,43 @@ package top.theillusivec4.curiouselytra.loader;
 
 import nerdhub.cardinal.components.api.event.ItemComponentCallbackV2;
 import net.fabricmc.api.ModInitializer;
-import net.minecraft.item.Items;
+import net.fabricmc.fabric.api.event.registry.RegistryEntryAddedCallback;
+import net.fabricmc.loader.api.FabricLoader;
+import net.minecraft.item.ElytraItem;
+import net.minecraft.item.Item;
+import net.minecraft.util.registry.Registry;
 import top.theillusivec4.curios.api.CuriosApi;
 import top.theillusivec4.curios.api.CuriosComponent;
 import top.theillusivec4.curios.api.SlotTypeInfo.BuildScheme;
 import top.theillusivec4.curios.api.SlotTypePreset;
 import top.theillusivec4.curiouselytra.core.CurioElytra;
+import top.theillusivec4.curiouselytra.loader.integration.NetheritePlusIntegration;
 
 public class CuriousElytraMod implements ModInitializer {
 
   @Override
   public void onInitialize() {
     CuriosApi.enqueueSlotType(BuildScheme.REGISTER, SlotTypePreset.BACK.getInfoBuilder().build());
-    ItemComponentCallbackV2.event(Items.ELYTRA).register(
+    boolean isNetheriteLoaded = FabricLoader.getInstance().isModLoaded("netherite_plus");
+
+    for (Item item : Registry.ITEM) {
+
+      if (item instanceof ElytraItem ||
+          (isNetheriteLoaded && NetheritePlusIntegration.isNetheriteElytra(item))) {
+        registerComponent(item);
+      }
+    }
+    RegistryEntryAddedCallback.event(Registry.ITEM).register((rawId, id, item) -> {
+
+      if (item instanceof ElytraItem ||
+          (isNetheriteLoaded && NetheritePlusIntegration.isNetheriteElytra(id))) {
+        registerComponent(item);
+      }
+    });
+  }
+
+  private static void registerComponent(Item itemIn) {
+    ItemComponentCallbackV2.event(itemIn).register(
         ((item, itemStack, componentContainer) -> componentContainer
             .put(CuriosComponent.ITEM, new CurioElytra(itemStack))));
   }